update toast messages to unified experience
